// src/enums/CategoryColors.ts
export enum CategoryColors {
	ACTINIDE = "tomato",
	ALKALI_METAL = "#6A1B9A",
	ALKALINE_EARTH_METAL = "#C62828",
	LANTHANIDE = "#AD1457",
	METALLOID = "#00695C",
	NOBLE_GAS = "#00838F",
	POST_TRANSITION_METAL = "#2E7D32",
	REACTIVE_NONMETAL = "#0277BD",
	TRANSITION_METAL = "#283593",
	UNKNOWN = "#9E9D24"
};

export default CategoryColors;